requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
pydantic-settings==2.7.1
httpx==0.28.1
pyyaml==6.0.2
thefuzz==0.22.1
python-levenshtein==0.26.1
